Read local natives in wtest npm command
